refactor: refactoring router
