--- src/App.css
@@ -1,10 +1,0 @@
-.main-content {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	margin-top: 6rem;
-	padding-top: 30px;
-	padding-bottom: 30px;
-	transition: all 300ms ease-in-out;
-	width: 100%;
-}
