+++ src/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import { RouterProvider } from "react-router-dom";
-import router from "./routes";
+import router from "@/app.routes";
 
 // TanStack Query 클라이언트 생성
 const queryClient = new QueryClient({
