+++ src/pages/control/control.routes.tsx
@@ -1,0 +1,12 @@
+import { Control, Participants } from "./pages";
+
+export const controlRoutes = [
+	{
+		path: "/control",
+		element: <Control />,
+	},
+	{
+		path: "/control/:questionSetId/participants",
+		element: <Participants />,
+	},
+];
