+++ src/pages/control/index.ts
@@ -1,4 +1,5 @@
 export * from "./components";
+export * from "./control.routes";
 export * from "./hooks";
 export * from "./layouts";
 export * from "./pages";
