+++ src/pages/creation/creation.routes.tsx
@@ -1,0 +1,8 @@
+import { QuestionCreation } from "./pages";
+
+export const creationRoutes = [
+	{
+		path: "/creation",
+		element: <QuestionCreation />,
+	},
+];
