+++ src/pages/creation/index.ts
@@ -1,4 +1,5 @@
 export * from "./components";
+export * from "./creation.routes";
 export * from "./hooks";
 export * from "./pages";
 export * from "./types";
