+++ src/pages/home/home.routes.tsx
@@ -1,0 +1,8 @@
+import Home from "./Home";
+
+export const homeRoutes = [
+	{
+		path: "/",
+		element: <Home />,
+	},
+];
