+++ src/pages/home/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Home";
+export * from "./home.routes";
